Don't overwrite mjkey.txt if it exists

+++ scripts/setup_linux.sh
@@ -183,7 +183,9 @@ if [[ "${_arg_modify_bashrc}" = on ]]; then
 fi
 
 # We need a MuJoCo key to import mujoco_py
-cp "${_arg_mjkey}" "${HOME}/.mujoco/mjkey.txt"
+if [[ ! -f "${HOME}/.mujoco/mjkey.txt" ]]; then
+  cp "${_arg_mjkey}" "${HOME}/.mujoco/mjkey.txt"
+fi
 
 if [[ "${_arg_modify_bashrc}" != on ]]; then
   echo -e "\nRemember to execute the following commands before running garage:"
